# LLM: Optimize transformer int4 loading