--- a/python/llm/src/bigdl/llm/transformers/convert.py
+++ b/python/llm/src/bigdl/llm/transformers/convert.py
@@ -49,8 +49,11 @@ def _replace_with_int4_linear(model, modules_to_not_convert=None,
         if param.data.device == torch.device('meta'):
             from accelerate.utils.modeling import set_module_tensor_to_device
             param = model_state_dict[name]
-            set_module_tensor_to_device(model, name, "cpu", torch.empty(*param.size(), dtype=torch.float32))
-   
+            set_module_tensor_to_device(model,
+                                        name,
+                                        "cpu",
+                                        torch.empty(*param.size(), dtype=torch.float32))
+
     for name, module in model.named_children():
         if current_key_name is None:
             current_key_name = []
@@ -107,6 +110,5 @@ def ggml_convert_int4(model, convert_shape_only=False):
             "an issue on github if you think this is a bug."
         )
     else:
-        # model.to(torch.float32)
-        pass
+        model.to(torch.float32)
     return model