## Description

In this PR we optimize transformer int4 model loading speed by:

1. using `low_cpu_mem_usage` to let huggingface original api to init model without init its weights at first, and also allocate the memory no more than  ~1x model size CPU memory as the official description introduces.
2. Directly pass an empty state_dict to from_pretained to avoid reading from local file at the first initialization(this weights will be abort)

Test shows that our loading benefits from this pr:

original vicuna 13B without low_cpu_mem_usage :55.64694786071777s

... with low_cpu_mem_usage: 9.221513748168945s

load and quantization: 32.92671084403992s

load int4 without optimization:  53.27334189414978s

load int4 with low_cpu_mem_usage: 7.347492694854736s

load int4 with low_cpu_mem_usage and without state_dict:  5.813234806060791s

Sanity Check:
```python

from bigdl.llm.transformers import AutoModelForCausalLM
from transformers import LlamaTokenizer

llama_model_path = "/home/ruonan/models/vicuna-13b/"
# llama_model_path = "/mnt/disk2/models/merged_gpt4all_llama_7b"
output_dir = "./vicuna-13b"

import time

start = time.time()
model = AutoModelForCausalLM.from_pretrained(llama_model_path,
                                             load_in_4bit=True)
during = time.time() - start
model.save_pretrained(output_dir)

start = time.time()
model = AutoModelForCausalLM.from_pretrained(output_dir)
during_0 = time.time() - start

tokenizer = LlamaTokenizer.from_pretrained(llama_model_path)
input = "what is AI?"
input_token = tokenizer(input, return_tensors="pt")
output_token = model.generate(**input_token)

summary = tokenizer.batch_decode(output_token,
                                 skip_special_tokens=True,
                                 spaces_between_special_tokens=False)
print(summary)

print(f"quan time {during}, load time {during_0}")
```
output:
```
Note: If there are warnings about mismatched during the loading process, please ignore them as it is part of the normal flow. The model will be reconverted to the format of BigDL after loading.
./vicuna-13b/pytorch_model.bin
/opt/anaconda3/envs/changmin-chatllm/lib/python3.9/site-packages/transformers/generation/utils.py:1313: UserWarning: Using `max_length`'s default (20) to control the generation length. This behaviour is deprecated and will be removed from the config in v5 of Transformers -- we recommend using `max_new_tokens` to control the maximum length of the generation.
  warnings.warn(
['what is AI?\n\nArtificial intelligence (AI) is the simulation of human']
```