@jason-dai 2023-07-11
> Test shows that our loading benefits from this pr:
> 
> load and quantization: 32.92671084403992s
> 
> load int4 without optimization: 53.27334189414978s
> 
> load int4 with low_cpu_mem_usage: 7.347492694854736s
> 
> load int4 with low_cpu_mem_usage and without state_dict: 5.813234806060791s

How about loading the original FP16 or FP32 model?